pyproject.toml
[project]
name = "randose"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.5.0",
    "einops>=0.8.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pytorch-msssim>=1.0.0",
    "simpleitk>=2.5.3",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "wandb>=0.24.2",
    "zetascale>=2.8.8",
]
